# src/SvarOMatic.Backend/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Swagger - Svar-o-Matic",
    "description" : "This is the backend for a solution that allows employees of Miles AS to efficiently match incoming project requests with existing consultant expertise.\n\nSome useful links:\n- [The Svar-o-Matic repository](https://github.com/miles-no/svar-o-matic)",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "drift@miles.no"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.11"
  },
...
  "servers" : [ {
    "url" : "http://localhost:5090"
  } ],
...
  "paths" : {
...
    "/krav/{consultantId}/{inquiryId}/add" : {
      "post" : {
        "tags" : [ "krav" ],
        "summary" : "Add a new krav to a forespørsel",
        "description" : "Add a new krav to a forespørsel",
        "operationId" : "createKrav",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "inquiryId",
          "in" : "path",
          "description" : "Id of the forespørsel that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "Data about the krav",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KravRequest"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/KravRequest"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/KravRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_inquiry_auth" : [ "write:inquiry", "read:inquiry" ]
        } ]
      }
    },
    "/krav/{consultantId}/{inquiryId}/{kravId}/edit" : {
      "post" : {
        "tags" : [ "krav" ],
        "summary" : "Edit an existing krav to a forespørsel",
        "description" : "Edit an existing krav to a forespørsel",
        "operationId" : "editKrav",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "inquiryId",
          "in" : "path",
          "description" : "Id of the forespørsel that the edit is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "kravId",
          "in" : "path",
          "description" : "Id of the krav that the edit is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "Data about the krav",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KravRequest"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/KravRequest"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/KravRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_inquiry_auth" : [ "write:inquiry", "read:inquiry" ]
        } ]
      }
    },
    "/krav/{consultantId}/{inquiryId}/{kravId}/delete" : {
      "post" : {
        "tags" : [ "krav" ],
        "summary" : "Delete a krav from a forespørsel",
        "description" : "Delete a krav from a forespørsel",
        "operationId" : "deleteKrav",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "inquiryId",
          "in" : "path",
          "description" : "Id of the forespørsel that the deletion is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "kravId",
          "in" : "path",
          "description" : "Id of the krav to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_inquiry_auth" : [ "write:inquiry", "read:inquiry" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "KravRequest" : {
        "type" : "object",
        "properties" : {
          "text" : {
            "type" : "string"
          },
          "isMustHave" : {
            "type" : "boolean"
          }
        },
        "xml" : {
          "name" : "##default"
        }
      },
...
      "ApiResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "code" : 0,
          "type" : "type",
          "message" : "message"
        },
        "xml" : {
          "name" : "##default"
        }
      }
    },
...
    "securitySchemes" : {
      "svaromatic_docs_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://coming-soon.swagger.io/oauth/authorize",
            "scopes" : {
              "write:cvs" : "modify docs for a consultant",
              "read:cvs" : "read your docs"
            }
          }
        }
      },
      "svaromatic_inquiry_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://coming-soon.swagger.io/oauth/authorize",
            "scopes" : {
              "write:cvs" : "modify inquiries for a consultant",
              "read:cvs" : "read inquiries"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}